// src/app/models/instruction_newinstruction.model.ts
export interface INewInstruction{
    branch_id?: string;
    inst_sl_no?: string;
    paye_cust_type?: string;
    paye_cust_grp_id?: string;
    paye_br_id?: string;
    paye_acc_no?: string;
    mail_addrs?: string;
    benf_type?: string;
    inst_status?: string;
    make_by?: string;
    inst_rule_id?: string;
    benf_cust_grp_id?: string;
    benf_prio_no?: string;
    benf_acc_no?: string;
    benf_br_id?: string;
    trm_tot_no?: string;
    trm_freq?: string;
    trm_instl_amt?: string;
    trm_tot_amt?: string;
    trm_tot_on?: string;
}