// src/app/models/instruction_getbeneficiary.model.ts
export interface IGetBeneficiaryResponse {
        benf_acc_no?: string;
        benf_acc_type?: string;
        rule_no?: string;
        trm_freq?: string;
        trm_instl_amt?: string;
        trm_tot_amt?: string;
        trm_tot_no?: string;
        trm_tot_on?: string;
        trm_instl_amt_paid?: string;
        lst_inst_pay_dt?: string;
        branchID?: string;
    }