// src/app/models/app_enum.model.ts
export enum ActivityType {
    LOGIN = 1,
    LOGOUT = 2,
    FUNDTRANSFEROWNBANK = 3,
    EDT = 4,
    ERROR = 5,
    SECURITYUPDATE = 6,
    FUNDTRANSFERFAILED = 7,
    ADD = 8,
    REGISTER = 9,
    CHECK = 10,
    AUTHORIZE = 11,
    DLT = 12,
    OPERATIONFAILED = 13,
    SIGNUPREQUESTSUCCESS = 14,
    SIGNUPREQUESTFAILED = 15,
    UNAUTHORIZEDACCESS = 16,
    INVALIDTPIN = 17,
    NOTIFICATION = 18,
    DEVLOG = 19,
    LOGINFAILED = 20,
    BILLPAY = 21,
    TOPUP = 22,
    QRPAY = 23,
    FUNDTRANSFEREFT = 24,
    FUNDTRANSFERRTGS = 25,
    FUNDTRANSFEROWNBANKOTHERACC = 26,
    TRPROFILECHANGE = 27,
    FUNDTRANSFERBKASH = 28,
    CREDITCARDBILLPAY = 29,
    QRCASH = 30,
    FUNDTRANSFERNAGAD = 31,
    FUNDTRANSFERNPSB = 32,
    FUNDTRANSFERSSLPAY = 33,
    BENEFICIARYADD = 34,
    BENEFICIARYUPDATE = 35,
    BENEFICIARYDELETE = 36,
    REQUEST = 37,
    UTILITYBILLPAYDHAKAWASA = 38,
    UTILITYBILLPAYNSU = 39,
    UTILITYBILLPAYDPDC = 40,
    UTILITYBILLPAYBRAC = 41,
    UTILITYBILLPAYDESCO = 42,
    UTILITYBILLPAYBGDCL = 43,
    UTILITYBILLPAYDOZE = 44,
    UTILITYBILLPAYBTCLMOTN = 45,
    UTILITYBILLPAYTITASGAS = 46,
    UTILITYBILLPAYTITASNONMETERGAS = 47,
    UTILITYBILLPAYSSLTOPUP = 48,
    UTILITYBILLPAYCREDITCARDBILLOWN = 49,
    UTILITYBILLPAYCREDITCARDBILLOTHER = 50,
    UTILITYBILLPAYNIDPAYMENT = 51,
    UTILITYBILLPAYSCHOOLBANKING = 52,
    ENCASHMENT = 53,
    BULKFUNDTRANSFER = 54,
    SCHEMEACCOUNTOPENING = 55,
    TIMEACCOUNTOPENING = 56,
    UTILITYBILLPAYDESCOPREPAIDOFFLINE = 57,
    UTILITYBILLPAYDESCOPREPAIDSMART = 58,
    UTILITYBILLPAYNESCOPOSTPAID = 59,
    UTILITYBILLPAYNESCOPREPAID = 60,
    UTILITYBILLPAYBREBPREPAID = 61,
    UTILITYBILLPAYBREBPOSTPAID = 62,
    UTILITYBILLPAYRAJWASA = 63,
    FUNDTRANSFERTAP = 64,
    FUNDTRANSFERUPAY = 65,
    FUNDTRANSFERROCKET = 66,
    UTILITYBILLPAYBREBPOSTICBS = 67
}

// Used in Account List for AccountType Parameter
export enum AccountType
{
    All = 0,
    Payable = 1, // Demand & Loan
    ACSummary = 2, // Demand, Time, Scheme
    Payee = 5,    // Only Demand
    FundTransfer = 13,
    OwnBankToAccount=17,
}